# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BlackBelt Commandos workforce operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestEvidenceUploadUrl
      tags: [operations]
      summary: Create a short-lived private evidence upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceUploadInput"
      responses:
        "200":
          description: Signed upload target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceUpload"
...
components:
...
  schemas:
...
    EvidenceUploadInput:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: number, minimum: 1 }
        contentType: { type: string }
    EvidenceUpload:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata: { type: object, additionalProperties: true }
...