# lib/api-zod/src/generated/types/evidenceUploadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BlackBelt Commandos workforce operations API
 * OpenAPI spec version: 0.1.0
 */

export interface EvidenceUploadInput {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: string;
}